+++ oidc_provider/admin.py
@@ -75,12 +75,16 @@ class ClientAdmin(admin.ModelAdmin):
 @admin.register(Code)
 class CodeAdmin(admin.ModelAdmin):
 
+    raw_id_fields = ['user']
+
     def has_add_permission(self, request):
         return False
 
 
 @admin.register(Token)
 class TokenAdmin(admin.ModelAdmin):
+
+    raw_id_fields = ['user']
 
     def has_add_permission(self, request):
         return False