+++ oidc_provider/lib/endpoints/token.py
@@ -1,11 +1,12 @@
-import inspect
-from base64 import urlsafe_b64encode
 import hashlib
+import inspect
 import logging
-from django.contrib.auth import authenticate
+from base64 import urlsafe_b64encode
 
+from django.contrib.auth import authenticate
 from django.http import JsonResponse
 
+from oidc_provider import settings
 from oidc_provider.lib.errors import (
     TokenError,
     UserAuthError,
@@ -17,14 +18,12 @@ from oidc_provider.lib.utils.token import (
     encode_id_token,
 )
 from oidc_provider.models import (
+    Client,
     Code,
     Token,
-    get_client_model
 )
-from oidc_provider import settings
 
 logger = logging.getLogger(__name__)
-Client = get_client_model()
 
 
 class TokenEndpoint(object):
@@ -77,7 +76,7 @@ class TokenEndpoint(object):
                 raise TokenError('invalid_grant')
 
             if not (self.code.client == self.client) \
-               or self.code.has_expired():
+                    or self.code.has_expired():
                 logger.debug('[Token] Invalid code: invalid client or code has expired')
                 raise TokenError('invalid_grant')
 
@@ -85,8 +84,8 @@ class TokenEndpoint(object):
             if self.params['code_verifier']:
                 if self.code.code_challenge_method == 'S256':
                     new_code_challenge = urlsafe_b64encode(
-                            hashlib.sha256(self.params['code_verifier'].encode('ascii')).digest()
-                        ).decode('utf-8').replace('=', '')
+                        hashlib.sha256(self.params['code_verifier'].encode('ascii')).digest()
+                    ).decode('utf-8').replace('=', '')
                 else:
                     new_code_challenge = self.params['code_verifier']
 
@@ -135,6 +134,27 @@ class TokenEndpoint(object):
         else:
             logger.debug('[Token] Invalid grant type: %s', self.params['grant_type'])
             raise TokenError('unsupported_grant_type')
+
+    def validate_requested_scopes(self):
+        """
+        Handling validation of requested scope for grant_type=[password|client_credentials]
+        """
+        token_scopes = []
+        if self.params['scope']:
+            # See https://tools.ietf.org/html/rfc6749#section-3.3
+            # The value of the scope parameter is expressed
+            # as a list of space-delimited, case-sensitive strings
+            for scope_requested in self.params['scope'].split(' '):
+                if scope_requested in self.client.scope:
+                    token_scopes.append(scope_requested)
+                else:
+                    logger.debug('[Token] The request scope %s is not supported by client %s',
+                                 scope_requested, self.client.client_id)
+                    raise TokenError('invalid_scope')
+        # if no scopes requested assign client's scopes
+        else:
+            token_scopes.extend(self.client.scope)
+        return token_scopes
 
     def create_response_dic(self):
         if self.params['grant_type'] == 'authorization_code':
@@ -231,11 +251,11 @@ class TokenEndpoint(object):
 
     def create_access_token_response_dic(self):
         # See https://tools.ietf.org/html/rfc6749#section-4.3
-
+        token_scopes = self.validate_requested_scopes()
         token = create_token(
             self.user,
             self.client,
-            self.params['scope'].split(' '))
+            token_scopes)
 
         id_token_dic = create_id_token(
             token=token,
@@ -256,15 +276,17 @@ class TokenEndpoint(object):
             'expires_in': settings.get('OIDC_TOKEN_EXPIRE'),
             'token_type': 'bearer',
             'id_token': encode_id_token(id_token_dic, token.client),
+            'scope': ' '.join(token.scope)
         }
 
     def create_client_credentials_response_dic(self):
         # See https://tools.ietf.org/html/rfc6749#section-4.4.3
+        token_scopes = self.validate_requested_scopes()
 
         token = create_token(
             user=None,
             client=self.client,
-            scope=self.client.scope)
+            scope=token_scopes)
 
         token.save()
 
@@ -272,7 +294,7 @@ class TokenEndpoint(object):
             'access_token': token.access_token,
             'expires_in': settings.get('OIDC_TOKEN_EXPIRE'),
             'token_type': 'bearer',
-            'scope': self.client._scope,
+            'scope': ' '.join(token.scope),
         }
 
     @classmethod