+++ oidc_provider/models.py
@@ -3,13 +3,10 @@ import binascii
 from hashlib import md5, sha256
 import json
 
-from django.apps import apps
 from django.db import models
 from django.utils import timezone
 from django.utils.translation import gettext_lazy as _
 from django.conf import settings
-
-from oidc_provider import settings as oidc_settings
 
 
 CLIENT_TYPE_CHOICES = [
@@ -56,13 +53,12 @@ class ResponseType(models.Model):
         return u'{0}'.format(self.description)
 
 
-class AbstractClient(models.Model):
+class Client(models.Model):
 
     name = models.CharField(max_length=100, default='', verbose_name=_(u'Name'))
     owner = models.ForeignKey(
         settings.AUTH_USER_MODEL, verbose_name=_(u'Owner'), blank=True,
-        null=True, default=None, on_delete=models.SET_NULL,
-        related_name='%(app_label)s_%(class)s_set')
+        null=True, default=None, on_delete=models.SET_NULL, related_name='oidc_clients_set')
     client_type = models.CharField(
         max_length=30,
         choices=CLIENT_TYPE_CHOICES,
@@ -72,8 +68,7 @@ class AbstractClient(models.Model):
                     u' of their credentials. <b>Public</b> clients are incapable.'))
     client_id = models.CharField(max_length=255, unique=True, verbose_name=_(u'Client ID'))
     client_secret = models.CharField(max_length=255, blank=True, verbose_name=_(u'Client SECRET'))
-    response_types = models.ManyToManyField(
-        ResponseType, related_name='%(app_label)s_%(class)s_set')
+    response_types = models.ManyToManyField(ResponseType)
     jwt_alg = models.CharField(
         max_length=10,
         choices=JWT_ALGS,
@@ -119,7 +114,6 @@ class AbstractClient(models.Model):
     class Meta:
         verbose_name = _(u'Client')
         verbose_name_plural = _(u'Clients')
-        abstract = True
 
     def __str__(self):
         return u'{0}'.format(self.name)
@@ -163,21 +157,9 @@ class AbstractClient(models.Model):
         return self.redirect_uris[0] if self.redirect_uris else ''
 
 
-class Client(AbstractClient):
-    class Meta(AbstractClient.Meta):
-        swappable = 'OIDC_CLIENT_MODEL'
-
-
-def get_client_model():
-    """ Return the Application model that is active in this project. """
-    return apps.get_model(oidc_settings.get('OIDC_CLIENT_MODEL'))
-
-
 class BaseCodeTokenModel(models.Model):
 
-    client = models.ForeignKey(
-        oidc_settings.get('OIDC_CLIENT_MODEL'), verbose_name=_(u'Client'),
-        on_delete=models.CASCADE)
+    client = models.ForeignKey(Client, verbose_name=_(u'Client'), on_delete=models.CASCADE)
     expires_at = models.DateTimeField(verbose_name=_(u'Expiration Date'))
     _scope = models.TextField(default='', verbose_name=_(u'Scopes'))